backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-pptx==1.0.2
sqlalchemy==2.0.36
python-multipart==0.0.20
aiofiles==24.1.0
google-genai>=1.0.0
python-dotenv>=1.0.0
